requirements.txt
Flask
flask-cors
gunicorn
numpy
Pillow 
BeautifulSoup4 
requests
 